refactored opc-client to use lib opc
