+++ lib/scala/opc/pom.xml
@@ -1,10 +1,10 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <groupId>com.hashmapinc.tempus.edge</groupId>
-  <artifactId>opc-client</artifactId>
-  <version>0.1.1</version>
+  <artifactId>opc</artifactId>
+  <version>0.1.0</version>
   <name>${project.artifactId}</name>
-  <description>Tempus edge application for converting iofog messages to opc writes and opc reads to iofog messages.</description>
+  <description>Tempus edge library for opc connectivity.</description>
   <inceptionYear>2018</inceptionYear>
 
   <properties>
@@ -22,6 +22,11 @@
       <version>${scala.version}</version>
     </dependency>
     <dependency>
+      <groupId>org.eclipse.milo</groupId>
+      <artifactId>sdk-client</artifactId>
+      <version>0.2.1</version>
+    </dependency>
+    <dependency>
       <groupId>com.typesafe.scala-logging</groupId>
       <artifactId>scala-logging_2.12</artifactId>
       <version>3.7.2</version>
@@ -37,14 +42,14 @@
       <version>0.1.0</version>
     </dependency>
     <dependency>
-      <groupId>com.hashmapinc.tempus.edge</groupId>
-      <artifactId>iofog</artifactId>
-      <version>0.1.0</version>
+      <groupId>org.bouncycastle</groupId>
+      <artifactId>bcprov-jdk15on</artifactId>
+      <version>1.58</version>
     </dependency>
     <dependency>
-      <groupId>com.hashmapinc.tempus.edge</groupId>
-      <artifactId>opc</artifactId>
-      <version>0.1.0</version>
+      <groupId>org.bouncycastle</groupId>
+      <artifactId>bcpkix-jdk15on</artifactId>
+      <version>1.58</version>
     </dependency>
 
     <!-- Test -->
@@ -132,71 +137,6 @@
         <configuration>
           <source>1.8</source>
           <target>1.8</target>
-        </configuration>
-      </plugin>
-      <!--==================================================================-->
-
-      <!--==================================================================-->
-      <!-- setup package to generate an uber-jar -->
-      <!--==================================================================-->
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-shade-plugin</artifactId>
-        <version>3.1.0</version>
-        <configuration>
-          <!-- put your configurations here -->
-          <createDependencyReducedPom>false</createDependencyReducedPom>
-          <finalName>${project.artifactId}-uber</finalName>
-          <transformers>
-            <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-              <mainClass>com.hashmapinc.tempus.edge.opcClient.Driver</mainClass>
-            </transformer>
-          </transformers>
-          <filters>
-            <filter>
-              <artifact>*:*</artifact>
-              <excludes>
-                <exclude>META-INF/license/**</exclude>
-                <exclude>META-INF/*</exclude>
-                <exclude>META-INF/maven/**</exclude>
-                <exclude>LICENSE</exclude>
-                <exclude>NOTICE</exclude>
-                <exclude>/*.txt</exclude>
-                <exclude>build.properties</exclude>
-              </excludes>
-            </filter>
-          </filters>
-        </configuration>
-        <executions>
-          <execution>
-            <phase>package</phase>
-            <goals>
-              <goal>shade</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
-      <!--==================================================================-->
-
-      <!--==================================================================-->
-      <!-- Configure dockerfile plugin -->
-      <!--==================================================================-->
-      <plugin>
-        <groupId>com.spotify</groupId>
-        <artifactId>dockerfile-maven-plugin</artifactId>
-        <version>1.3.6</version>
-        <executions>
-          <execution>
-            <id>default</id>
-            <goals>
-              <goal>build</goal>
-              <goal>push</goal>
-            </goals>
-          </execution>
-        </executions>
-        <configuration>
-          <repository>hashmapinc/tempus-edge-${project.artifactId}</repository>
-          <tag>${project.version}</tag>
         </configuration>
       </plugin>
       <!--==================================================================-->
