+++ lib/scala/opc/src/main/scala/com/hashmapinc/tempus/edge/opc/OpcSecurity.scala
@@ -1,4 +1,4 @@
-package com.hashmapinc.tempus.edge.opcClient.opc
+package com.hashmapinc.tempus.edge.opc
 
 import java.io.FileInputStream
 import java.security.{KeyPair, KeyPairGenerator}
@@ -14,7 +14,6 @@ import org.eclipse.milo.opcua.stack.core.util.SelfSignedCertificateBuilder
 import org.eclipse.milo.opcua.stack.core.util.SelfSignedCertificateGenerator
 
 import com.hashmapinc.tempus.edge.proto.OpcConfig
-import com.hashmapinc.tempus.edge.opcClient.Config
 
 object OpcSecurity {
   private val log = Logger(getClass())
@@ -23,13 +22,13 @@ object OpcSecurity {
   // Generate keypair and certificate at run time
   val clientKeyPair = SelfSignedCertificateGenerator.generateRsaKeyPair(2048)
   val clientCertificate = (new SelfSignedCertificateBuilder(clientKeyPair)
-    .setCommonName("Tempus Edge OPC Client")
+    .setCommonName("Tempus Edge OPC Library")
     .setOrganization("hashmapinc")
     .setOrganizationalUnit("tempus edge")
     .setLocalityName("Atlanta")
     .setStateName("GA")
     .setCountryCode("US")
-    .setApplicationUri("urn:hashmapinc:tempus:edge:opc-client")
+    .setApplicationUri("urn:hashmapinc:tempus:edge:opc")
   ).build
   
   /**
